# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TraclyTag Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/my-company:
    get:
      operationId: getMyCompany
      tags: [companies]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMyCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                companyUrl:
                  type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies/my-company/regenerate-api-key:
    post:
      operationId: regenerateCompanyApiKey
      tags: [companies]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies/{id}:
    delete:
      operationId: deleteCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    put:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"


  # ---------- USERS ----------
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- LOCATIONS ----------
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- BATCHES ----------
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [batches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    delete:
      operationId: deleteBatch
      tags: [batches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- CODES ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: ["string", "null"]
          pattern: "^[0-9]{2}[a-zA-Z0-9]{10}[a-zA-Z0-9][zZ][a-zA-Z0-9]?$"
        companyUrl:
          type: ["string", "null"]
        apiKey:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        cin:
          type: ["string", "null"]
        msmeRegistrationNo:
          type: ["string", "null"]
        fssaiLicenseNo:
          type: ["string", "null"]
        drugLicenseNo:
          type: ["string", "null"]
        iecCode:
          type: ["string", "null"]
        companyPrefix:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, address, gstin, createdAt]
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: ["string", "null"]
          pattern: "^[0-9]{2}[a-zA-Z0-9]{10}[a-zA-Z0-9][zZ][a-zA-Z0-9]?$"
        companyUrl:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        cin:
          type: ["string", "null"]
        msmeRegistrationNo:
          type: ["string", "null"]
        fssaiLicenseNo:
          type: ["string", "null"]
        drugLicenseNo:
          type: ["string", "null"]
        iecCode:
          type: ["string", "null"]
        companyPrefix:
          type: ["string", "null"]
      required: [name, email, address]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        skuId:
          type: string
        name:
          type: string
        skuSize:
          type: string
        marketedBy:
          type: string
        sapDescription:
          type: ["string", "null"]
        gtin:
          type: ["string", "null"]
        mrp:
          type: number
        registrationNo:
          type: ["string", "null"]
        hsnCode:
          type: ["string", "null"]
        gstRate:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        weightValue:
          type: ["number", "null"]
        weightUnit:
          type: ["string", "null"]
        packagingType:
          type: ["string", "null"]
        shelfLifeDays:
          type: ["integer", "null"]
        countryOfOrigin:
          type: ["string", "null"]
        isGs1Compliant:
          type: ["boolean", "null"]
        l1Size:
          type: integer
          description: Units per 1st level inner pack
        l2Size:
          type: integer
          description: 1st level packs per 2nd level inner pack
        shipperSize:
          type: integer
          description: 2nd level packs per shipper
        cautionLogoUrl:
          type: ["string", "null"]
        productLogoUrl:
          type: ["string", "null"]
        labelPdfUrl:
          type: ["string", "null"]
        expiryDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - skuId
        - name
        - skuSize
        - marketedBy
        - mrp
        - l1Size
        - l2Size
        - shipperSize
        - expiryDate
        - createdAt
    CreateProductBody:
      type: object
      properties:
        skuId:
          type: string
        name:
          type: string
        skuSize:
          type: string
        marketedBy:
          type: string
        sapDescription:
          type: ["string", "null"]
        gtin:
          type: ["string", "null"]
          description: 14 digit GTIN
        mrp:
          type: number
        registrationNo:
          type: ["string", "null"]
        hsnCode:
          type: ["string", "null"]
        gstRate:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        weightValue:
          type: ["number", "null"]
        weightUnit:
          type: ["string", "null"]
        packagingType:
          type: ["string", "null"]
        shelfLifeDays:
          type: ["integer", "null"]
        countryOfOrigin:
          type: ["string", "null"]
        isGs1Compliant:
          type: ["boolean", "null"]
        l1Size:
          type: integer
        l2Size:
          type: integer
        shipperSize:
          type: integer
        cautionLogoUrl:
          type: ["string", "null"]
        productLogoUrl:
          type: ["string", "null"]
        labelPdfUrl:
          type: ["string", "null"]
        expiryDate:
          type: string
          format: date
      required:
        - skuId
        - name
        - skuSize
        - marketedBy
        - mrp
        - l1Size
        - l2Size
        - shipperSize
        - expiryDate
    Location:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        locationType:
          type: string
        uniqueName:
          type: string
        locationName:
          type: string
        contactNo:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
        gln:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - locationType
        - uniqueName
        - locationName
        - contactNo
        - state
        - city
        - address
        - createdAt
    CreateLocationBody:
      type: object
      properties:
        locationType:
          type: string
        uniqueName:
          type: string
        locationName:
          type: string
        contactNo:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
        gln:
          type: ["string", "null"]
      required:
        - locationType
        - uniqueName
        - locationName
        - contactNo
        - state
        - city
        - address
    Batch:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        batchNumber:
          type: string
        mfgDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - productName
        - batchNumber
        - mfgDate
        - expiryDate
        - createdAt
    CreateBatchBody:
      type: object
      properties:
        productId:
          type: integer
        batchNumber:
          type: string
        mfgDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
      required: [productId, batchNumber, mfgDate, expiryDate]
...